pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backend-django"
version = "1.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django==5.0.0",
    "django-cors-headers==4.3.1",
    "django-ninja==1.3.0",
    "django-ninja-jwt==5.3.5",
    "duckdb==1.4.1",
    "email-validator==2.1.0",
    "openpyxl==3.1.5",
    "pandas==2.2.0",
    "psycopg2-binary==2.9.9",
    "pydantic==2.5.0",
    "pyjwt==2.10.1",
    "python-dateutil==2.9.0",
    "python-dotenv==1.0.0",
]

[tool.setuptools]
packages = ["study", "common", "config", "report", "project", "studies", "claudedocs"]

[project.optional-dependencies]
dev = [
    "sphinx",
    "numpydoc",
    "pydata-sphinx-theme",
    "sphinx-copybutton",
    "sphinx-design",
    "ruff",
    "mypy",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
exclude = [
    ".*/migrations/.*",
    "manage.py",
    "tests/.*",
    "scripts/.*",
]

[[tool.mypy.overrides]]
module = [
    "duckdb.*",
    "psycopg2.*",
    "studies.config",
    "studies.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "study.models",
    "report.models",
    "project.models",
    "common.models",
]
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "duckdb.*",
    "psycopg2.*",
    "studies.config",
]
ignore_missing_imports = true
